Implement multi-server support

+++ cogs/expandochannels.py
@@ -13,8 +13,11 @@ class ExpandoChannels():
         if ctx.message.author.voice is not None:
             if ctx.message.author.voice.channel is not None:
                 if ctx.message.author.voice.channel.name[0] == "🎮":
-                    await ctx.message.author.voice.channel.edit(name="🎮 " + ctx.arg)
-    
+                    if self.bot.config[ctx.guild.id]["expando_channels"]:
+                        await ctx.message.author.voice.channel.edit(name="🎮 " + ctx.arg)
+                    else :
+                        await ctx.send("Your admins have disabled Expando Channels. Get them to enable it to use this feature!")
+
     async def update_empty_channel(self, starting_channel):
         #Go through expando channels starting with the channel that was triggered
         chan_index = starting_channel.guild.voice_channels.index(starting_channel)
@@ -37,10 +40,12 @@ class ExpandoChannels():
 
         if before.channel is not None:
             if len(before.channel.members) == 0 and before.channel.name[0] == "🎮":
-                await before.channel.delete()
+                if self.bot.config[before.channel.guild.id]["expando_channels"]:
+                    await before.channel.delete()
         if after.channel is not None:
             if len(after.channel.members) == 1 and after.channel.name[0] == "🎮":
-                await self.update_empty_channel(after.channel)
+                if self.bot.config[after.channels.guild.id]["expando_channels"]:
+                    await self.update_empty_channel(after.channel)
 
 def setup(bot : commands.Bot):
     print("setting up expandochannels")
